Restructure fields into own directory
- Put fields in own files in order to make it easier to make more variations on field
- Change tests to use new components
